backend/requirements.txt
fastapi
uvicorn
python-dotenv
protobuf
sqlalchemy
python-multipart
pytest
httpx